+++ ios/THEOplayerRCTViewManager.swift
@@ -2,9 +2,6 @@
 //  THEOplayerRCTViewManager.swift
 //  Theoplayer
 //
-//  Created by William van Haevre on 14/01/2022.
-//  Copyright © 2022 Facebook. All rights reserved.
-//
 
 import Foundation
 import UIKit