+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.12.1] - 23-09-14
+
+### Fixed
+
+- Fixed an issue on Android where the player SDK dependency could resolve to version 6.+.
+- Fixed an issue on iOS where the player integration dependencies could resolve to version 6.+.
+
 ## [2.12.0] - 23-09-04
 
 ### Added